+++ packages/select/index.js
@@ -11,6 +11,7 @@ import { when } from 'lit/directives/when.js';
 import { activateI18n } from '../i18n';
 import { kebabCaseAttributes } from '../utils';
 
+import { messages as daMessages } from './locales/da/messages.mjs';
 import { messages as enMessages } from './locales/en/messages.mjs';
 import { messages as fiMessages } from './locales/fi/messages.mjs';
 import { messages as nbMessages } from './locales/nb/messages.mjs';
@@ -36,31 +37,51 @@ export class WarpSelect extends kebabCaseAttributes(WarpElement) {
     // Whether to show optional text
     optional: { type: Boolean, reflect: true },
 
+    // Renders the field in a disabled state.
+    disabled: { type: Boolean, reflect: true },
+
+    // Renders the field in a readonly state.
+    readOnly: { type: Boolean, relfect: true },
+
     _options: { state: true },
   };
 
   static styles = [WarpElement.styles];
 
+  constructor() {
+    super();
+    activateI18n(enMessages, nbMessages, fiMessages, daMessages);
+
+    this._options = this.innerHTML;
+  }
+
+  firstUpdated() {
+    // autofocus doesn't seem to behave properly in Safari and FireFox, therefore we set the focus here:
+    this.autoFocus && this.shadowRoot.querySelector('select').focus();
+  }
+
+  handleKeyDown(event) {
+    if (this.readOnly && (event.key === ' ' || event.key === 'ArrowDown' || event.key === 'ArrowUp')) {
+      event.preventDefault();
+    }
+  }
+
   get #classes() {
-    return classNames({
-      [ccSelect.default]: true,
-      [ccSelect.invalid]: this.invalid,
-    });
+    return classNames([
+      ccSelect.base,
+      !this.invalid && !this.disabled && !this.readOnly && ccSelect.default,
+      this.invalid && ccSelect.invalid,
+      this.disabled && ccSelect.disabled,
+      this.readOnly && ccSelect.readOnly,
+    ]);
   }
 
   get #helpTextClasses() {
-    return classNames({
-      [ccHelpText.helpText]: true,
-      [ccHelpText.helpTextColor]: !this.invalid,
-      [ccHelpText.helpTextColorInvalid]: this.invalid,
-    });
+    return classNames([ccHelpText.base, this.invalid ? ccHelpText.colorInvalid : ccHelpText.color]);
   }
 
   get #chevronClasses() {
-    return classNames({
-      [ccSelect.chevron]: true,
-      [ccSelect.chevronDisabled]: this.disabled,
-    });
+    return classNames([ccSelect.chevron, this.disabled && ccSelect.chevronDisabled]);
   }
 
   get #id() {
@@ -71,19 +92,12 @@ export class WarpSelect extends kebabCaseAttributes(WarpElement) {
     return this.hint ? `${this.#id}__hint` : undefined;
   }
 
-  constructor() {
-    super();
-    activateI18n(enMessages, nbMessages, fiMessages);
-
-    this._options = this.innerHTML;
-  }
-
   render() {
     return html`<div class="${ccSelect.wrapper}">
       ${when(
         this.label,
         () =>
-          html`<label class="${ccLabel.label}" for="${this.#id}">
+          html`<label class="${ccLabel.base}" for="${this.#id}">
             ${this.label}
             ${when(
               this.optional,
@@ -102,10 +116,11 @@ export class WarpSelect extends kebabCaseAttributes(WarpElement) {
         <select
           class="${this.#classes}"
           id="${this.#id}"
-          ?autofocus=${this.autoFocus}
+          ?disabled=${this.disabled}
           aria-describedby="${ifDefined(this.#helpId)}"
           aria-invalid="${ifDefined(this.invalid)}"
-          aria-errormessage="${ifDefined(this.invalid && this.#helpId)}">
+          aria-errormessage="${ifDefined(this.invalid && this.#helpId)}"
+          @keydown=${this.handleKeyDown}>
           ${unsafeHTML(this._options)}
         </select>
         <div class="${this.#chevronClasses}">